Refactor asset ordering in the API
Use database sort to sort assets. Use title for default asset sort. The
"assets" variable was getting passed around in assetmgr/views.py, but
ultimately ignored by the AssetResource.render_list() function. This
function can actually use the list of assets and preserve ordering.

This fixes the bugs we were seeing with asset ordering.
